:art: Refactor to use already implemented CustomDebugHotkeysMixin.
