+++ api/src/main/java/com/noxcrew/noxesium/api/util/DebugOption.java
@@ -1,8 +1,8 @@
-package com.noxcrew.noxesium.feature.dkeys;
+package com.noxcrew.noxesium.api.util;
 
 import org.jetbrains.annotations.Nullable;
 
-public enum DebugKey {
+public enum DebugOption {
     TOGGLE_PROFILER(49, null),
     TOGGLE_FPS_CHARTS(50, null),
     TOGGLE_NETWORK_CHARTS(51, null),
@@ -19,13 +19,14 @@ public enum DebugKey {
     HELP(81, "debug.help.help"),
     DUMP_TEXTURES(83, "debug.dump_dynamic_textures.help"),
     RELOAD_RESOURCES(84, "debug.reload_resourcepacks.help"),
+    NOXESIUM_SETTINGS(86, "debug.noxesium_settings.help"),
     ESCAPE(256, "debug.pause.help"),
     GAME_MODE_SWITCHER(293, "debug.gamemodes.help");
 
     private final int keyCode;
     private final @Nullable String translationKey;
 
-    DebugKey(int keyCode, @Nullable String helpTranslationKey) {
+    DebugOption(int keyCode, @Nullable String helpTranslationKey) {
         this.keyCode = keyCode;
         this.translationKey = helpTranslationKey;
     }
@@ -38,8 +39,8 @@ public enum DebugKey {
         return translationKey;
     }
 
-    public static @Nullable DebugKey getByTranslationKey(String translationKey) {
-        for (DebugKey key : values()) {
+    public static @Nullable DebugOption getByTranslationKey(String translationKey) {
+        for (DebugOption key : values()) {
             if (key.translationKey != null && key.translationKey.equals(translationKey)) {
                 return key;
             }
@@ -47,8 +48,8 @@ public enum DebugKey {
         return null;
     }
 
-    public static @Nullable DebugKey getByKeyCode(int keyCode) {
-        for (DebugKey key : values()) {
+    public static @Nullable DebugOption getByKeyCode(int keyCode) {
+        for (DebugOption key : values()) {
             if (key.keyCode == keyCode) {
                 return key;
             }
