+++ common/src/main/resources/assets/noxesium/lang/en_us.json
@@ -1,5 +1,8 @@
 {
   "debug.noxesium_settings.help": "F3 + V = Open Noxesium settings menu",
+  "debug.warning.option.disabled": "This debug option is disabled.",
+  "debug.warning.option.disabled_by_server": "The server you are playing on has disabled this debug option.",
+
 
   "debug.fps_overlay": "%s fps",
   "debug.game_time_overlay": "%s st, %s gt",
@@ -42,9 +45,6 @@
   "noxesium.options.enum.bottom": "Bottom Corner",
   "noxesium.options.enum.top_flipped": "Opposite Top Corner",
   "noxesium.options.enum.bottom_flipped": "Opposite Bottom Corner",
-
-  "noxesium.warning.debug_option.disabled": "This debug option is disabled.",
-  "noxesium.warning.debug_option.disabled_by_server": "The server you are playing on has disabled this debug option.",
 
   "itemGroup.noxesium.server_items": "Custom Blocks",
