+++ common/src/main/resources/noxesium-common.mixins.json
@@ -28,7 +28,6 @@
     "rules.HandItemOverrideAttackFix",
     "rules.HandItemOverrideInventoryMixin",
     "rules.HandItemOverrideRenderMixin",
-    "rules.KeyboardHandlerMixin",
     "rules.OptionInstanceMixin",
     "rules.SpinAttackCollisionMixin",
     "rules.entity.DebugRendererMixin",
